src/requirements.txt
python-dotenv
msal
requests